i am able find the solution for the adding fast item by a merchant

+++ lib/features/merchant/domain/entities/daily_aggregate_entity.dart
@@ -1,7 +1,8 @@
 /// Domain Entity - Aggregated Item
 class AggregatedItemEntity {
   final String name;
-  final int quantity;
+  final double
+  quantity; // Changed from int to double to support fractional quantities
   final double revenue;
 
   const AggregatedItemEntity({
